code rearranged to have the dedicated storage driver for both dataverses and datasets (but not datafiles). #6872
